feat: run statistic *extraction* (not post-process for now) in parallel
