+++ src/cli/benchmark-app.ts
@@ -59,7 +59,7 @@ async function benchmark() {
 	const verboseAdd = options.verbose ? ['--verbose'] : []
 
 	const pool = new LimitBenchmarkPool(
-		`${__dirname}/../cli/benchmark-helper-app`,
+		`${__dirname}/benchmark-helper-app`,
 		files.map(f => [f.content, '--output', options.output, '--slice', options.slice, ...verboseAdd]),
 		limit,
 		options.parallel
